Extracting BaseClients
